# specs/003-hierarquia-catalogo-herois/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Darkest Dungeon - Catálogo de Heróis
  version: 0.1.0
  description: |
    Contratos HTTP para a hierarquia de entidades (Ser, Habilidade, Item) e o catálogo
    oficial das 20 classes de heróis. Todas as mensagens de erro seguem PT-BR.
servers:
  - url: /
paths:
  /classes:
    get:
      summary: Listar todas as classes de herói
      operationId: listarClasses
      responses:
        "200":
          description: Lista das 20 classes cadastradas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClasseResumo"
  /classes/{id}:
    get:
      summary: Consultar uma classe pelo id
      operationId: obterClasse
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Classe encontrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClasseDetalhe"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /classes/{id}/habilidades:
    get:
      summary: Listar habilidades de herói disponíveis para a classe
      operationId: listarHabilidadesDaClasse
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Habilidades associadas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HabilidadeResumo"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /habilidades:
    get:
      summary: Listar habilidades cadastradas
      operationId: listarHabilidades
      parameters:
        - in: query
          name: categoria
          schema:
            $ref: "#/components/schemas/CategoriaDeHabilidade"
        - in: query
          name: classe
          schema:
            $ref: "#/components/schemas/ClasseDeHeroi"
      responses:
        "200":
          description: Lista de habilidades.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HabilidadeResumo"
  /habilidades/{id}:
    get:
      summary: Consultar uma habilidade pelo id
      operationId: obterHabilidade
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Habilidade encontrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HabilidadeDetalhe"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /habilidades/combate:
    post:
      summary: Cadastrar habilidade de combate
      operationId: criarHabilidadeDeCombate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarHabilidadeDeCombate"
      responses:
        "201":
          description: Habilidade de combate cadastrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HabilidadeDetalhe"
        "400":
          $ref: "#/components/responses/EntradaInvalida"
  /habilidades/acampamento:
    post:
      summary: Cadastrar habilidade de acampamento
      operationId: criarHabilidadeDeAcampamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarHabilidadeDeAcampamento"
      responses:
        "201":
          description: Habilidade de acampamento cadastrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HabilidadeDetalhe"
        "400":
          $ref: "#/components/responses/EntradaInvalida"
  /habilidades/inimigo:
    post:
      summary: Cadastrar habilidade de inimigo
      operationId: criarHabilidadeDeInimigo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarHabilidadeDeInimigo"
      responses:
        "201":
          description: Habilidade de inimigo cadastrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HabilidadeDetalhe"
        "400":
          $ref: "#/components/responses/EntradaInvalida"
  /personagens:
    post:
      summary: Criar um Personagem baseado em uma Classe
      operationId: criarPersonagem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarPersonagem"
      responses:
        "201":
          description: Personagem criado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonagemDetalhe"
        "400":
          $ref: "#/components/responses/EntradaInvalida"
  /personagens/{id}:
    get:
      summary: Consultar Personagem
      operationId: obterPersonagem
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Personagem encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonagemDetalhe"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /personagens/{id}/equipar:
    post:
      summary: Equipar itens ao Personagem
      operationId: equiparPersonagem
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquiparPersonagem"
      responses:
        "200":
          description: Equipamento aplicado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonagemDetalhe"
        "400":
          $ref: "#/components/responses/EntradaInvalida"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /inimigos:
    post:
      summary: Cadastrar Inimigo
      operationId: criarInimigo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarInimigo"
      responses:
        "201":
          description: Inimigo cadastrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InimigoDetalhe"
        "400":
          $ref: "#/components/responses/EntradaInvalida"
  /inimigos/{id}:
    get:
      summary: Consultar Inimigo
      operationId: obterInimigo
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Inimigo encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InimigoDetalhe"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /armas:
    post:
      summary: Cadastrar Arma
      operationId: criarArma
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarArma"
      responses:
        "201":
          description: Arma cadastrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemDetalhe"
        "400":
          $ref: "#/components/responses/EntradaInvalida"
  /armaduras:
    post:
      summary: Cadastrar Armadura
      operationId: criarArmadura
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarArmadura"
      responses:
        "201":
          description: Armadura cadastrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemDetalhe"
        "400":
          $ref: "#/components/responses/EntradaInvalida"
  /acessorios:
    post:
      summary: Cadastrar Acessorio
      operationId: criarAcessorio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarAcessorio"
      responses:
        "201":
          description: Acessorio cadastrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemDetalhe"
        "400":
          $ref: "#/components/responses/EntradaInvalida"
  /itens/{id}:
    get:
      summary: Consultar Item (Arma, Armadura ou Acessorio)
      operationId: obterItem
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Item encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemDetalhe"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /mapa-de-cobertura:
    get:
      summary: Consultar o Mapa de Cobertura completo do catálogo
      operationId: obterMapaDeCobertura
      responses:
        "200":
          description: Mapa completo, agrupado por classe.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MapaDeCoberturaPorClasse"
  /mapa-de-cobertura/{classeId}:
    get:
      summary: Consultar o Mapa de Cobertura de uma classe específica
      operationId: obterMapaDeCoberturaPorClasse
      parameters:
        - in: path
          name: classeId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Mapa da classe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapaDeCoberturaPorClasse"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
  responses:
    NaoEncontrado:
      description: Recurso não encontrado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErroResponse"
    EntradaInvalida:
      description: Entrada inválida.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErroResponse"
  schemas:
    ErroResponse:
      type: object
      required: [mensagem]
      properties:
        mensagem:
          type: string
          description: Mensagem descritiva em PT-BR.
        campo:
          type: string
          nullable: true
    ClasseDeHeroi:
      type: string
      enum:
        - Abominacao
        - Antiquario
        - Besteiro
        - CacadorDeRecompensas
        - Cruzado
        - LadraoDeCova
        - BoboDaCorte
        - MestreDeCaca
        - Leproso
        - Infernal
        - Bandido
        - Musqueteiro
        - Veterano
        - Ocultista
        - MedicoDaPeste
        - Vestal
        - Flagelante
        - Rompedor
        - Duelista
        - Fugitivo
    TipoDeInimigo:
      type: string
      enum: [Humano, Besta, Profano, Sobrenatural, Vampirico, Casca, Rochoso]
    CategoriaDeHabilidade:
      type: string
      enum: [Combate, Acampamento, Inimigo]
    AlvoDeEfeito:
      type: string
      enum: [Self, Aliado, Inimigo]
    UnidadeDeEfeito:
      type: string
      enum: [Percentual, Pontos, Rodadas]
    AlvoDeAcampamento:
      type: string
      enum: [Self, UmAliado, TodosOsAliados, PartyInteira]
    RaridadeDeAcessorio:
      type: string
      enum: [Comum, Incomum, Rara, MuitoRara, CrimsonCourt, Crystalline, Set]
    EscopoDeLimite:
      type: string
      enum: [Batalha, Acampamento]
    SinalDeEfeito:
      type: string
      enum: [Positivo, Negativo]
    CategoriaDeCobertura:
      type: string
      enum: [HabilidadeCombate, HabilidadeAcampamento, ResistenciaBase]
    EstadoDeAtributo:
      type: string
      enum: [Coletado, Pendente, NaoAplicavel]
    ResistenciasBase:
      type: object
      required:
        - atordoamento
        - sangramento
        - envenenamento
        - debuff
        - movimento
        - doenca
        - golpeMortal
        - armadilha
      properties:
        atordoamento: { type: number, minimum: 0, maximum: 100 }
        sangramento: { type: number, minimum: 0, maximum: 100 }
        envenenamento: { type: number, minimum: 0, maximum: 100 }
        debuff: { type: number, minimum: 0, maximum: 100 }
        movimento: { type: number, minimum: 0, maximum: 100 }
        doenca: { type: number, minimum: 0, maximum: 100 }
        golpeMortal: { type: number, minimum: 0, maximum: 100 }
        armadilha: { type: number, minimum: 0, maximum: 100 }
    ClasseResumo:
      type: object
      required: [id, classe, nomeExibicao]
      properties:
        id: { type: string, format: uuid }
        classe: { $ref: "#/components/schemas/ClasseDeHeroi" }
        nomeExibicao: { type: string }
        nomeOriginal: { type: string }
    ClasseDetalhe:
      allOf:
        - $ref: "#/components/schemas/ClasseResumo"
        - type: object
          required: [resistenciasBase]
          properties:
            resistenciasBase:
              $ref: "#/components/schemas/ResistenciasBase"
    EfeitoDeHabilidade:
      type: object
      required: [nomeDoEfeito, alvo, valor, unidade, chanceBase]
      properties:
        nomeDoEfeito: { type: string, maxLength: 60 }
        alvo: { $ref: "#/components/schemas/AlvoDeEfeito" }
        valor: { type: number }
        unidade: { $ref: "#/components/schemas/UnidadeDeEfeito" }
        duracaoEmRodadas:
          type: integer
          nullable: true
          minimum: 0
          maximum: 20
        chanceBase: { type: number, minimum: 0, maximum: 100 }
    LimitePorUso:
      type: object
      nullable: true
      required: [escopo, maximoUsos]
      properties:
        escopo: { $ref: "#/components/schemas/EscopoDeLimite" }
        maximoUsos: { type: integer, minimum: 1, maximum: 10 }
    HabilidadeResumo:
      type: object
      required: [id, categoria, nomeExibicao]
      properties:
        id: { type: string, format: uuid }
        categoria: { $ref: "#/components/schemas/CategoriaDeHabilidade" }
        nomeExibicao: { type: string }
        nomeOriginal: { type: string }
    HabilidadeDetalhe:
      type: object
      required: [id, categoria, nomeExibicao, descricao, efeitos]
      properties:
        id: { type: string, format: uuid }
        categoria: { $ref: "#/components/schemas/CategoriaDeHabilidade" }
        nomeExibicao: { type: string }
        nomeOriginal: { type: string }
        descricao: { type: string }
        classes:
          type: array
          items: { $ref: "#/components/schemas/ClasseDeHeroi" }
        efeitos:
          type: array
          items: { $ref: "#/components/schemas/EfeitoDeHabilidade" }
        limitePorUso: { $ref: "#/components/schemas/LimitePorUso" }
        combate:
          type: object
          nullable: true
          properties:
            posicoesValidas:
              type: array
              items: { type: integer, minimum: 1, maximum: 4 }
            posicoesQueAtinge:
              type: array
              items: { type: integer, minimum: 1, maximum: 4 }
            alvoEmArea: { type: boolean }
            modificadorDano: { type: number }
            modificadorAcerto: { type: number }
            modificadorCritico: { type: number }
        acampamento:
          type: object
          nullable: true
          properties:
            custoDeDescanso: { type: integer, minimum: 0, maximum: 20 }
            alvo: { $ref: "#/components/schemas/AlvoDeAcampamento" }
        inimigo:
          type: object
          nullable: true
          properties:
            condicaoDeAparecer: { type: string, nullable: true }
            chanceDeExecucao: { type: number, nullable: true, minimum: 0, maximum: 100 }
    CriarHabilidadeDeCombate:
      type: object
      required: [nomeExibicao, nomeOriginal, descricao, classes, posicoesValidas, posicoesQueAtinge, modificadorDano, modificadorAcerto, modificadorCritico, efeitos]
      properties:
        nomeExibicao: { type: string, maxLength: 80 }
        nomeOriginal: { type: string, maxLength: 80 }
        descricao: { type: string, maxLength: 400 }
        classes:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/ClasseDeHeroi" }
        posicoesValidas:
          type: array
          minItems: 1
          items: { type: integer, minimum: 1, maximum: 4 }
        posicoesQueAtinge:
          type: array
          minItems: 1
          items: { type: integer, minimum: 1, maximum: 4 }
        alvoEmArea: { type: boolean, default: false }
        modificadorDano: { type: number }
        modificadorAcerto: { type: number }
        modificadorCritico: { type: number }
        efeitos:
          type: array
          items: { $ref: "#/components/schemas/EfeitoDeHabilidade" }
        limitePorUso: { $ref: "#/components/schemas/LimitePorUso" }
    CriarHabilidadeDeAcampamento:
      type: object
      required: [nomeExibicao, nomeOriginal, descricao, classes, custoDeDescanso, alvo, efeitos]
      properties:
        nomeExibicao: { type: string, maxLength: 80 }
        nomeOriginal: { type: string, maxLength: 80 }
        descricao: { type: string, maxLength: 400 }
        classes:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/ClasseDeHeroi" }
        custoDeDescanso: { type: integer, minimum: 0, maximum: 20 }
        alvo: { $ref: "#/components/schemas/AlvoDeAcampamento" }
        efeitos:
          type: array
          items: { $ref: "#/components/schemas/EfeitoDeHabilidade" }
        limitePorUso: { $ref: "#/components/schemas/LimitePorUso" }
    CriarHabilidadeDeInimigo:
      type: object
      required: [nomeExibicao, nomeOriginal, descricao, efeitos]
      properties:
        nomeExibicao: { type: string, maxLength: 80 }
        nomeOriginal: { type: string, maxLength: 80 }
        descricao: { type: string, maxLength: 400 }
        condicaoDeAparecer: { type: string, nullable: true, maxLength: 200 }
        chanceDeExecucao: { type: number, nullable: true, minimum: 0, maximum: 100 }
        efeitos:
          type: array
          items: { $ref: "#/components/schemas/EfeitoDeHabilidade" }
    NivelDeArma:
      type: object
      required: [nivel, danoMinimo, danoMaximo, critico, velocidade]
      properties:
        nivel: { type: integer, minimum: 1, maximum: 5 }
        danoMinimo: { type: integer, minimum: 0 }
        danoMaximo: { type: integer, minimum: 0 }
        critico: { type: number, minimum: 0, maximum: 100 }
        velocidade: { type: integer, minimum: 0 }
    NivelDeArmadura:
      type: object
      required: [nivel, hpAdicional, esquiva]
      properties:
        nivel: { type: integer, minimum: 1, maximum: 5 }
        hpAdicional: { type: integer, minimum: 0 }
        esquiva: { type: number, minimum: 0, maximum: 100 }
    EfeitoDeAcessorio:
      type: object
      required: [nome, valor, unidade, sinal]
      properties:
        nome: { type: string, maxLength: 80 }
        valor: { type: number }
        unidade: { $ref: "#/components/schemas/UnidadeDeEfeito" }
        sinal: { $ref: "#/components/schemas/SinalDeEfeito" }
    ItemDetalhe:
      type: object
      required: [id, tipo, nomeExibicao]
      properties:
        id: { type: string, format: uuid }
        tipo:
          type: string
          enum: [Arma, Armadura, Acessorio]
        nomeExibicao: { type: string }
        nomeOriginal: { type: string }
        descricao: { type: string }
        classeElegivel: { $ref: "#/components/schemas/ClasseDeHeroi" }
        niveisArma:
          type: array
          nullable: true
          items: { $ref: "#/components/schemas/NivelDeArma" }
        niveisArmadura:
          type: array
          nullable: true
          items: { $ref: "#/components/schemas/NivelDeArmadura" }
        raridade: { $ref: "#/components/schemas/RaridadeDeAcessorio" }
        classeExclusiva:
          allOf:
            - $ref: "#/components/schemas/ClasseDeHeroi"
          nullable: true
        conjuntoId: { type: string, format: uuid, nullable: true }
        efeitosAcessorio:
          type: array
          nullable: true
          items: { $ref: "#/components/schemas/EfeitoDeAcessorio" }
    CriarArma:
      type: object
      required: [nomeExibicao, nomeOriginal, descricao, classeElegivel, niveis]
      properties:
        nomeExibicao: { type: string, maxLength: 80 }
        nomeOriginal: { type: string, maxLength: 80 }
        descricao: { type: string, maxLength: 400 }
        classeElegivel: { $ref: "#/components/schemas/ClasseDeHeroi" }
        niveis:
          type: array
          minItems: 5
          maxItems: 5
          items: { $ref: "#/components/schemas/NivelDeArma" }
    CriarArmadura:
      type: object
      required: [nomeExibicao, nomeOriginal, descricao, classeElegivel, niveis]
      properties:
        nomeExibicao: { type: string, maxLength: 80 }
        nomeOriginal: { type: string, maxLength: 80 }
        descricao: { type: string, maxLength: 400 }
        classeElegivel: { $ref: "#/components/schemas/ClasseDeHeroi" }
        niveis:
          type: array
          minItems: 5
          maxItems: 5
          items: { $ref: "#/components/schemas/NivelDeArmadura" }
    CriarAcessorio:
      type: object
      required: [nomeExibicao, nomeOriginal, descricao, raridade, efeitos]
      properties:
        nomeExibicao: { type: string, maxLength: 80 }
        nomeOriginal: { type: string, maxLength: 80 }
        descricao: { type: string, maxLength: 400 }
        raridade: { $ref: "#/components/schemas/RaridadeDeAcessorio" }
        classeExclusiva:
          allOf:
            - $ref: "#/components/schemas/ClasseDeHeroi"
          nullable: true
        conjuntoId:
          type: string
          format: uuid
          nullable: true
          description: Metadata para agrupamento de conjunto de Acessórios. O cálculo do bônus quando duas peças do mesmo conjunto são equipadas está fora do escopo desta feature.
        efeitos:
          type: array
          items: { $ref: "#/components/schemas/EfeitoDeAcessorio" }
    CriarPersonagem:
      type: object
      required: [nome, classe]
      description: |
        Cria um Personagem. `habilidadesEquipadas` MUST conter apenas Ids de
        `HabilidadeDeHeroi` associadas à Classe informada (ver associação Classe × Habilidade)
        e MUST respeitar o limite de 6 habilidades de Combate + 6 habilidades de Acampamento
        (a categorização é conferida pelo serviço via discriminator da `Habilidade`).
        Violações retornam 400 com o schema `ErroResponse` em PT-BR.
      properties:
        nome: { type: string, minLength: 1, maxLength: 60 }
        classe: { $ref: "#/components/schemas/ClasseDeHeroi" }
        habilidadesEquipadas:
          type: array
          items:
            type: string
            format: uuid
    EquiparPersonagem:
      type: object
      properties:
        armaId: { type: string, format: uuid, nullable: true }
        armaduraId: { type: string, format: uuid, nullable: true }
        acessoriosIds:
          type: array
          maxItems: 2
          items: { type: string, format: uuid }
    HabilidadeDePersonagem:
      type: object
      required: [habilidadeId, habilitada, treinada, equipada]
      properties:
        habilidadeId: { type: string, format: uuid }
        habilitada: { type: boolean }
        treinada: { type: boolean }
        equipada: { type: boolean }
    PersonagemDetalhe:
      type: object
      required: [id, nome, classe, resistencias]
      properties:
        id: { type: string, format: uuid }
        nome: { type: string }
        classe: { $ref: "#/components/schemas/ClasseDeHeroi" }
        stress: { type: integer, minimum: 0, maximum: 200 }
        chanceDeVirtude: { type: integer, minimum: 0, maximum: 100 }
        aflicao: { type: string, nullable: true }
        virtude: { type: string, nullable: true }
        estadoPortasDaMorte: { type: boolean }
        recuperouPortasDaMorte: { type: boolean }
        recuperouAtaqueCardiaco: { type: boolean }
        resistencias:
          $ref: "#/components/schemas/ResistenciasBase"
        habilidades:
          type: array
          items: { $ref: "#/components/schemas/HabilidadeDePersonagem" }
        armaEquipadaId: { type: string, format: uuid, nullable: true }
        armaduraEquipadaId: { type: string, format: uuid, nullable: true }
        acessoriosEquipadosIds:
          type: array
          items: { type: string, format: uuid }
        inventario:
          type: array
          minItems: 4
          maxItems: 4
          items:
            type: string
            format: uuid
            nullable: true
    CriarInimigo:
      type: object
      required: [nome, tipo]
      properties:
        nome: { type: string, minLength: 1, maxLength: 60 }
        tipo: { $ref: "#/components/schemas/TipoDeInimigo" }
        habilidadesIds:
          type: array
          items: { type: string, format: uuid }
    InimigoDetalhe:
      type: object
      required: [id, nome, tipo]
      properties:
        id: { type: string, format: uuid }
        nome: { type: string }
        tipo: { $ref: "#/components/schemas/TipoDeInimigo" }
        habilidades:
          type: array
          items:
            $ref: "#/components/schemas/HabilidadeResumo"
    EntradaCobertura:
      type: object
      required: [categoria, chaveDoAtributo, estado]
      properties:
        categoria: { $ref: "#/components/schemas/CategoriaDeCobertura" }
        chaveDoAtributo: { type: string }
        estado: { $ref: "#/components/schemas/EstadoDeAtributo" }
        notas: { type: string, nullable: true }
    MapaDeCoberturaPorClasse:
      type: object
      required: [classeId, classe, entradas]
      properties:
        classeId: { type: string, format: uuid }
        classe: { $ref: "#/components/schemas/ClasseDeHeroi" }
        entradas:
          type: array
          items: { $ref: "#/components/schemas/EntradaCobertura" }