# src/DarkestDungeon.Domain/Seres/TipoDeInimigo.cs
namespace DarkestDungeon.Domain.Seres;

/// Categoria estrutural de um Inimigo (usada para modificadores contextuais no jogo).
public enum TipoDeInimigo
{
    Humano,
    Besta,
    Profano,
    Sobrenatural,
    Vampirico,
    Casca,
    Rochoso,
}